bank view

+++ remunapp/visualizer/debugger.rs
@@ -7,7 +7,7 @@ const NR_ROWS: usize = 40;
 pub struct Debugger {
     following_pc: bool,
     disassembly: Vec<(u16, Instruction)>,
-    cursor: u16,
+    cursor: u64,
     line_number: usize,
 }
 
@@ -20,17 +20,17 @@ impl Debugger {
     pub fn update(&mut self, ctx: &egui::Context, ui: &mut egui::Ui, state: &mut State) {
         ui.toggle_value(&mut self.following_pc, "Following PC");
         if self.following_pc {
-            self.cursor = state.pc;
+            self.cursor = state.pc as u64;
         }
         let input = ctx.input(|i| i.clone());
         crate::visualizer::scroll_area(ctx, ui, &mut self.line_number, &mut self.cursor);
         if input.key_pressed(egui::Key::Enter) || self.following_pc {
-            if let Some(ln) = self.disassembly.iter().position(|(addr, _)| *addr >= self.cursor) {
+            if let Some(ln) = self.disassembly.iter().position(|(addr, _)| *addr >= (self.cursor as u16)) {
                 self.line_number = ln;
             }
         }
         self.disassembly[self.line_number..(self.line_number+NR_ROWS)].iter().for_each(|(addr, i)| {
-            if self.cursor >= *addr && self.cursor <= *addr + (i.1.arity() as u16) {
+            if (self.cursor as u16) >= *addr && (self.cursor as u16) <= *addr + (i.1.arity() as u16) {
                 ui.label(RichText::new(format!("{addr:04X}: {i}")).color(Color32::GREEN));
             } else if state.pc >= *addr && state.pc <= *addr + (i.1.arity() as u16) {
                 ui.label(RichText::new(format!("{addr:04X}: {i}")).color(Color32::YELLOW));
